add ocr workflow edit
