+++ Kitodo/src/main/resources/messages/messages_de.properties
@@ -247,6 +247,7 @@ createNewspaperProcessesTask=Anlegen von Zeitungsvorg\u00E4ngen
 createNewBatchFromSelectedProcesses=Neuen Batch aus den gew\u00E4hlten Vorg\u00E4ngen erzeugen
 createNewDocket=Neuen Laufzettel anlegen
 createNewLdapGroup=Neue LDAP-Gruppe anlegen
+createNewOCRWorkflow=Neuen OCR Workflow anlegen
 createNewProcess=Einen neuen Vorgang anlegen
 createNewProcessFromTemplate=Einen neuen Vorgang auf Basis der Produktionsvorlage anlegen
 createNewRuleset=Neuen Regelsatz anlegen
@@ -1117,6 +1118,7 @@ editClient=Mandant bearbeiten
 editDocket=Laufzettel bearbeiten
 editLdapGroup=LDAP-Gruppe bearbeiten
 editLdapServer=LDAP-Server bearbeiten
+editOCRWorkflow=OCR Workflow bearbeiten
 editProcess=Vorgang bearbeiten
 editProcessMetaData=Metadaten des Vorgangs bearbeiten
 editProcessStructureData=Strukturdaten des Vorgangs bearbeiten
@@ -1161,6 +1163,7 @@ viewDocket=Laufzettel anzeigen
 viewLdapGroup=LDAP-Gruppe anzeigen
 viewLdapServer=LDAP-Server anzeigen
 viewMigration=Migrationsseite anzeigen
+viewOCRWorkflow=OCR Workflow anzeigen
 viewProcess=Vorgang anzeigen
 viewProcessMetaData=Metadaten des Vorgangs anzeigen
 viewProcessStructureData=Strukturdaten des Vorgangs anzeigen
