+++ Kitodo/src/main/resources/messages/messages_en.properties
@@ -252,6 +252,7 @@ createNewspaperProcessesTask=Creation of processes for a newspaper
 createNewBatchFromSelectedProcesses=create new batch of the selected processes
 createNewDocket=Create new docket
 createNewLdapGroup=Create new LDAP group
+createNewOCRWorkflow=Create OCR workflow
 createNewProcess=Create new process
 createNewProcessFromTemplate=Create process from template
 createNewRuleset=Create new rule set
@@ -1131,6 +1132,7 @@ editClient=Edit client
 editDocket=Edit docket
 editLdapGroup=Edit LDAP group
 editLdapServer=Edit LDAP-server
+editOCRWorkflow=Edit OCR workflow
 editProcess=Edit process
 editProject=Edit project
 editProperty=Edit property
@@ -1166,6 +1168,7 @@ viewProcess=View process
 viewProcessImages=View images of the process
 viewProcessMetaData=View metadata of the process
 viewProcessStructureData=View structure data of the process
+viewOCRWorkflow=View OCR workflow
 viewProject=View project
 viewRole=View role
 viewRuleset=View ruleset
