+++ Kitodo/src/main/webapp/pages/ocrWorkflowEdit.xhtml
@@ -19,41 +19,38 @@
         xmlns:p="http://primefaces.org/ui">
     <f:metadata>
         <f:viewParam name="id"/>
-        <f:viewAction action="#{RulesetForm.load(id)}"/>
+        <f:viewAction action="#{OCRWorkflowEditView.load(id)}"/>
     </f:metadata>
 
     <!--@elvariable id="isEditMode" type="boolean"-->
     <ui:param name="isEditMode"
-              value="#{RulesetForm.ruleset.id ne null and SecurityAccessController.hasAuthorityToEditRuleset()}"/>
+              value="#{OCRWorkflowEditView.ocrWorkflow.id ne null and SecurityAccessController.hasAuthorityToEditOCRWorkflow()}"/>
     <!--@elvariable id="isCreateMode" type="boolean"-->
     <ui:param name="isCreateMode"
-              value="#{RulesetForm.ruleset.id eq null and SecurityAccessController.hasAuthorityToAddRuleset()}"/>
+              value="#{OCRWorkflowEditView.ocrWorkflow.id eq null and SecurityAccessController.hasAuthorityToAddOCRWorkflow()}"/>
     <!--@elvariable id="isViewMode" type="boolean"-->
     <ui:param name="isViewMode"
-              value="#{not isEditMode and not isCreateMode and SecurityAccessController.hasAuthorityToViewRuleset()}"/>
+              value="#{not isEditMode and not isCreateMode and SecurityAccessController.hasAuthorityToViewOCRWorkflow()}"/>
 
     <ui:define name="contentHeader">
         <h:outputScript name="js/save.js" target="body" />
-        <script>
-            window.onload = function () {checkForm('editForm')};
-        </script>
         <h3 id="headerText">
-            <h:outputText value="#{msgs.createNewRuleset}" rendered="#{isCreateMode}"/>
-            <h:outputText value="#{msgs.editRuleset} (#{RulesetForm.ruleset.title})" rendered="#{isEditMode}"/>
-            <h:outputText value="#{msgs.viewRuleset} (#{RulesetForm.ruleset.title})" rendered="#{isViewMode}"/>
+            <h:outputText value="#{msgs.createNewOCRWorkflow}" rendered="#{isCreateMode}"/>
+            <h:outputText value="#{msgs.editOCRWorkflow} (#{OCRWorkflowEditView.ocrWorkflow.title})" rendered="#{isEditMode}"/>
+            <h:outputText value="#{msgs.viewOCRWorkflow} (#{OCRWorkflowEditView.ocrWorkflow.title})" rendered="#{isViewMode}"/>
         </h3>
         <p:commandButton id="save"
                          widgetVar="save"
                          value="#{msgs.save}"
-                         action="#{RulesetForm.save}"
+                         action="#{OCRWorkflowEditView.save}"
                          update="notifications"
                          icon="fa fa-floppy-o fa-lg"
                          iconPos="right"
                          rendered="#{isEditMode || isCreateMode}"
                          onclick="setConfirmUnload(false);PF('notifications').renderMessage({'summary':'#{msgs.rulesetSaving}','detail':'#{msgs.youWillBeRedirected}','severity':'info'});"
-                         disabled="#{RulesetForm.saveDisabled}"/>
+                         disabled="#{OCRWorkflowEditView.saveDisabled}"/>
         <p:commandButton id="saveButtonToggler"
-                         actionListener="#{RulesetForm.setSaveDisabled(false)}"
+                         actionListener="#{OCRWorkflowEditView.setSaveDisabled(false)}"
                          update="save"
                          style="display:none;"/>
         <p:button value="#{msgs.cancel}"
@@ -65,31 +62,23 @@
     </ui:define>
 
     <ui:define name="pageTabView">
-        <p:tabView id="rulesetTabView">
-            <p:tab id="rulesetTab" title="#{msgs.details}">
+        <p:tabView id="ocrWorkflowView">
+            <p:tab id="ocrWorkflowTab" title="#{msgs.details}">
                 <p:panelGrid columns="2" layout="grid">
                     <p:row>
                         <div>
                             <p:outputLabel for="title" value="#{msgs.title}"/>
                             <p:inputText id="title" class="input" placeholder="#{msgs.title}"
-                                         value="#{RulesetForm.ruleset.title}" disabled="#{isViewMode}"
+                                         value="#{OCRWorkflowEditView.ocrWorkflow.title}" disabled="#{isViewMode}"
                                          required="#{empty param['editForm:saveButtonToggler']}"/>
                         </div>
                         <div>
                             <p:outputLabel for="file" value="#{msgs.file}"/>
-                            <p:selectOneMenu id="file" value="#{RulesetForm.ruleset.file}" disabled="#{isViewMode}">
-                                <f:selectItems value="#{RulesetForm.rulesetFilenames}" var="rulesetFile" itemValue="#{rulesetFile}"
-                                               itemLabel="#{rulesetFile}"/>
+                            <p:selectOneMenu id="file" value="#{OCRWorkflowEditView.ocrWorkflow.file}" disabled="#{isViewMode}">
+                                <f:selectItems value="#{OCRWorkflowEditView.OCRWorkflowFilenames}" var="ocrWorkflowFile" itemValue="#{ocrWorkflowFile}"
+                                               itemLabel="#{ocrWorkflowFile}"/>
                                 <p:ajax event="change" oncomplete="toggleSave()"/>
                             </p:selectOneMenu>
-                        </div>
-                    </p:row>
-                    <p:row>
-                        <div>
-                            <p:outputLabel for="sortByRuleset" value="#{msgs.metadataSortByRuleset}"/>
-                            <p:selectBooleanCheckbox id="sortByRuleset" styleClass="switch input" disabled="#{isViewMode}"
-                                                     value="#{RulesetForm.ruleset.orderMetadataByRuleset}"
-                                                     onchange="toggleSave()"/>
                         </div>
                     </p:row>
                 </p:panelGrid>
@@ -100,9 +89,9 @@
     <ui:define name="breadcrumbs">
         <p:menuitem value="#{msgs.desktop}" url="desktop.jsf" icon="fa fa-home"/>
         <p:menuitem value="#{msgs.projects}" url="projects.jsf" icon="fa fa-archive"/>
-        <p:menuitem value="#{msgs.createNewRuleset}" rendered="#{isCreateMode}" icon="fa fa-file-code-o"/>
-        <p:menuitem value="#{msgs.editRuleset}" rendered="#{isEditMode}" icon="fa fa-file-code-o"/>
-        <p:menuitem value="#{msgs.viewRuleset}" rendered="#{isViewMode}" icon="fa fa-file-code-o"/>
+        <p:menuitem value="#{msgs.createNewOCRWorkflow}" rendered="#{isCreateMode}" icon="fa fa-file-code-o"/>
+        <p:menuitem value="#{msgs.editOCRWorkflow}" rendered="#{isEditMode}" icon="fa fa-file-code-o"/>
+        <p:menuitem value="#{msgs.viewOCRWorkflow}" rendered="#{isViewMode}" icon="fa fa-file-code-o"/>
     </ui:define>
 
 </ui:composition>
